remove ESC handling during completion
previously, pressing arrow keys right after completion "ate" the beginning of
the escape sequence and dumped the rest of the sequence into the input buffer.

personally, i do that way more often than cancel completions, so the tradeoff
makes sense here.  this could instead be done with a timeout, to tell if the
ESC was entered manually, but that's hacky and complex to do properly.